Add useImageRegistry hook and integrate in controllers
Introduced a new useImageRegistry hook to centralize image registry data fetching and fallback logic. Updated ChaosStudioClone, ChaosStudioEdit, and StudioOverview components to use this hook, replacing direct queries and improving code reuse.
