+++ chaoscenter/web/src/controllers/ChaosStudioClone/ChaosStudioClone.tsx
@@ -8,7 +8,7 @@ import { listExperiment, runChaosExperiment, saveChaosExperiment } from '@api/co
 import experimentYamlService from '@services/experiment';
 import { InfrastructureType } from '@api/entities';
 import Loader from '@components/Loader';
-import { useSearchParams, useUpdateSearchParams } from '@hooks';
+import { useSearchParams, useUpdateSearchParams, useImageRegistry } from '@hooks';
 import { ExperimentManifest, StudioMode } from '@models';
 
 export default function ChaosStudioCloneController(): React.ReactElement {
@@ -17,6 +17,7 @@ export default function ChaosStudioCloneController(): React.ReactElement {
   const searchParams = useSearchParams();
   const updateSearchParams = useUpdateSearchParams();
   const hasUnsavedChangesInURL = searchParams.get('unsavedChanges') === 'true';
+  const { imageRegistry } = useImageRegistry();
 
   // <!-- counting state since we have 2 async functions and need to flip state when both of said functions have resolved their promises -->
   const [showStudio, setShowStudio] = React.useState<number>(0);
@@ -56,7 +57,8 @@ export default function ChaosStudioCloneController(): React.ReactElement {
             id: experimentData?.infra?.infraID,
             namespace: experimentData.infra?.infraNamespace,
             environmentID: experimentData.infra?.environmentID
-          }
+          },
+          imageRegistry: imageRegistry
         })
         .then(() => setShowStudio(oldState => oldState + 1));
       const parsedManifest = parse(experimentData.experimentManifest) as ExperimentManifest;
@@ -66,7 +68,7 @@ export default function ChaosStudioCloneController(): React.ReactElement {
         .then(() => setShowStudio(oldState => oldState + 1));
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [experimentData, experimentID, hasUnsavedChangesInURL]);
+  }, [experimentData, experimentID, hasUnsavedChangesInURL, imageRegistry]);
 
   const [saveChaosExperimentMutation, { loading: saveChaosExperimentLoading }] = saveChaosExperiment({
     onError: error => showError(error.message)
